Solace/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.1
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
gunicorn==20.1.0
inflection==0.5.1
Pillow==9.2.0
psycopg2==2.9.5
pycparser==2.21
pytz==2022.5
sqlparse==0.4.3
whitenoise==6.2.0
